fix MediaStore.MediaColumns.DATA is deprecated

+++ app/src/main/java/com/example/imgur/data/helpers/ImagesPath.kt
@@ -2,6 +2,7 @@ package com.example.imgur.data.helpers
 
 import android.annotation.SuppressLint
 import android.content.ContentResolver
+import android.content.ContentUris
 import android.provider.MediaStore
 import java.util.*
 
@@ -12,19 +13,32 @@ object ImagesPath {
      *
      * @return the paths of images.
      */
+
     @SuppressLint("Recycle")
     fun getAllImagesPath(contentResolver: ContentResolver): ArrayList<String> {
-        val uri = MediaStore.Images.Media.EXTERNAL_CONTENT_URI
-        val projection = arrayOf(MediaStore.MediaColumns.DATA)
-        val cursor = contentResolver.query(uri, projection, null, null, null)
-        val columnIndex = cursor!!.getColumnIndexOrThrow(MediaStore.MediaColumns.DATA)
+        val externalUri = MediaStore.Images.Media.EXTERNAL_CONTENT_URI
         val listOfAllImages = ArrayList<String>()
         val absolutePathOfImage = StringBuilder()
-        while (cursor.moveToNext()) {
-            absolutePathOfImage.append(cursor.getString(columnIndex))
-            listOfAllImages.add(absolutePathOfImage.toString())
-            absolutePathOfImage.delete(0, absolutePathOfImage.capacity())
+
+        val cursor = contentResolver
+            .query(
+                externalUri,
+                arrayOf(MediaStore.MediaColumns._ID, MediaStore.MediaColumns.DATE_MODIFIED),
+                null,
+                null,
+                "${MediaStore.MediaColumns.DATE_MODIFIED} DESC"
+            )
+
+        if (cursor != null) {
+            while (cursor.moveToNext()) {
+                val id = cursor.getString(cursor.getColumnIndex(MediaStore.MediaColumns._ID))
+                val uri = ContentUris.withAppendedId(externalUri, id.toLong())
+                absolutePathOfImage.append(uri.toString())
+                listOfAllImages.add(absolutePathOfImage.toString())
+                absolutePathOfImage.delete(0, absolutePathOfImage.capacity())
+            }
         }
+        cursor?.close()
         return listOfAllImages
     }
 }
